# feat: improve asset service performance