Implementation for https://github.com/shapeshift/lib/issues/374

To summarize, the code provided in the issue description **greatly improves** the performance of the asset service, the gain is more than 2500 ms of execution time.

Some performance statistics :  (FCP is the value returned by reportWebVitals in the app)

- With legacy code : 
          FCP = **6560 ms** on average for 5 executions
          Function execution time = **2690 ms** on average for 5 executions
          
- With optimized code :
          FCP = **4050ms** on average for 5 executions
          Function execution time = **8 ms** on average for 5 executions

I saw no obvious other optimizations in the asset service, notably flattenAssetData function is not costly.