@cjthompson 2022-02-16
For the bounty, were you able to identify any other areas that could use performance improvements?