@chxash 2022-02-16
> For the bounty, were you able to identify any other areas that could use performance improvements?

I spent some time trying to find other pieces of code that could be improved, but did not find anything obvious.